Move case utils functions to a dedicated file
