+++ packages/atomic/src/components/commerce/atomic-product-multi-value-text/atomic-product-multi-value-text.ts
@@ -13,8 +13,8 @@ import {bindings} from '@/src/decorators/bindings';
 import {errorGuard} from '@/src/decorators/error-guard';
 import type {InitializableComponent} from '@/src/decorators/types';
 import {withTailwindStyles} from '@/src/decorators/with-tailwind-styles';
-import {getFieldValueCaption} from '../../../utils/field-utils';
-import {titleToKebab} from '../../../utils/utils';
+import {titleToKebab} from '@/src/utils/case-utils';
+import {getFieldValueCaption} from '@/src/utils/field-utils';
 import type {CommerceBindings} from '../atomic-commerce-interface/atomic-commerce-interface';
 
 /**
